+++ Test/other_variants/Lab_7/var_11/frand.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <stdio.h>
+#include <stdlib.h>
+
+int round(float f);
+float frand_n(int n);
+float normal_frand_n(int count, int n);
