+++ Test/other_variants/Lab_7/var_11/main.cpp
@@ -2,33 +2,11 @@
 // на интервале [-5., 5.] (использовать rand()). Подсчитать частоты
 // попадания в каждый из интервалов длиной 1 при выборке 500000
 
-#include <stdio.h>
-#include <stdlib.h>
+#include "frand.h"
 
 #define N 5
 #define NUM 500000
 #define POWER 12
-
-#define frand() ((float)(rand())/(RAND_MAX))
-
-// [-n ... n]
-float frand_n(int n) {
-  return frand() * n * 2 - n;
-}
-
-// [-n ... n] норм. распределение
-// Центральная Предельная теорема
-float normal_frand_n(int count, int n) {
-  float sum = 0;
-  for (int i = 0; i < count; i++)
-    sum += frand_n(n);
-  return sum / count;
-}
-
-// округление: i <= f < i+1
-int round(float f) { 
-  return (int)(f > 0 ? f : f - 1);
-}
 
 int main() {
   int freq[N * 2] = {0};
